Move integration tests out to `tests`
