+++ pgdo-test/src/lib.rs
@@ -3,10 +3,15 @@ use quote::quote;
 
 use pgdo::runtime::strategy::{Strategy, StrategyLike};
 
-/// Attribute macro to generate a test function for each runtime discovered by
-/// `pgdo`. The function will be named after the original function, with a
-/// suffix containing the version number. The test body itself will have a
-/// "magic" `runtime` variable available, which is a [`pgdo::runtime::Runtime`].
+/// Attribute macro to generate a test function for each discovered runtime.
+///
+/// The function will be named after the original function, with a suffix
+/// containing the version number. The test body itself will have a "magic"
+/// `runtime` variable available, which is a [`pgdo::runtime::Runtime`].
+///
+/// **Note** that this macro will NOT work in unit tests of `pgdo-lib` because
+/// it does not know that it's called `pgdo`. In any case, this macro is
+/// intended for integration tests, so this limitation seems reasonable.
 #[proc_macro_attribute]
 pub fn for_all_runtimes(_attr: TokenStream, item: TokenStream) -> TokenStream {
     // Parse the input tokens into a syntax tree.
@@ -34,7 +39,7 @@ pub fn for_all_runtimes(_attr: TokenStream, item: TokenStream) -> TokenStream {
             let expanded = quote! {
                 #[::std::prelude::v1::test]
                 #vis #signature_with_version {
-                    let runtime = crate::runtime::Runtime::new(#bindir)?;
+                    let runtime = ::pgdo::runtime::Runtime::new(#bindir)?;
                     #block
                 }
             };
